simplify solvers (#84)
* add objective and constraints in arguments of search()
* replace setValueFromDict with setValuesFromArray in Solution
* add callback in BaseSearch
