--- flopt/solvers/optuna_searches/base_optuna.py
@@ -1,83 +1,0 @@
-import timeout_decorator
-
-import flopt
-from flopt.solvers.base import BaseSearch
-from flopt.constants import VariableType, ExpressionType, SolverTerminateState
-from flopt.env import setup_logger
-
-
-logger = setup_logger(__name__)
-
-
-class OptunaSearch(BaseSearch):
-    """
-    Optuna Update
-      It has a incumbent solution anytime
-      1. Generate a new solution using Optuna sampler
-      2. Check a new solution can be incumbent solutions
-      3. Update incumbent solution
-
-    Parameters
-    ----------
-    n_trial : int
-        number of trials
-    """
-
-    name = "OptunaSearch(base)"
-    can_solve_problems = {
-        "Variable": VariableType.Number,
-        "Objective": ExpressionType.Any,
-        "Constraint": ExpressionType.Non,
-    }
-
-    def __init__(self):
-        super().__init__()
-        self.n_trial = 1e100
-
-    def createStudy(self):
-        raise NotImplementedError()
-
-    def search(self):
-        self.start_build()
-
-        self.createStudy()
-
-        self.end_build()
-
-        search_timelimit = self.timelimit - self.build_time
-
-        @timeout_decorator.timeout(search_timelimit, timeout_exception=TimeoutError)
-        def optimize():
-            self.study.optimize(self.objective, self.n_trial, timeout=search_timelimit)
-
-        optimize()
-
-        return SolverTerminateState.Normal
-
-    def objective(self, trial):
-        # set value into self.solution
-        self.trial_ix += 1
-        for var in self.solution:
-            if var.type() == VariableType.Binary:
-                var.setValue(trial.suggest_int(var.name, 0, 1))
-            elif var.type() == VariableType.Spin:
-                var.toBinary()
-                var.binary.setValue(trial.suggest_int(var.name, 0, 1))
-            lb = var.getLb(must_number=True)
-            ub = var.getUb(must_number=True)
-            if var.type() == VariableType.Integer:
-                var.setValue(trial.suggest_int(var.name, lb, ub))
-            elif var.type() == VariableType.Continuous:
-                var.setValue(trial.suggest_uniform(var.name, lb, ub))
-
-        # get objective value by self.solution
-        obj_value = self.getObjValue(self.solution)
-
-        # if solution is better thatn incumbent, then update best solution
-        self.registerSolution(self.solution, obj_value)
-
-        # callback
-        for callback in self.callbacks:
-            callback([self.solution], self.best_solution, self.best_obj_value)
-
-        return obj_value
