+++ flopt/solvers/random_search.py
@@ -7,16 +7,7 @@ logger = setup_logger(__name__)
 
 
 class RandomSearch(BaseSearch):
-    """
-    Random Sampling Search
-
-    It is a simple serach, as follows.
-
-    .. code-block:: python
-
-      def setNewSolution(self, *args, **kwargs):
-          self.solution.setRandom()
-    """
+    """Random Sampling Search"""
 
     name = "RandomSearch"
     can_solve_problems = {
@@ -29,21 +20,16 @@ class RandomSearch(BaseSearch):
         super().__init__()
         self.n_trial = 1e100
 
-    def search(self):
-        """
-        search a better solution using `self.setNewSolution()` function
-        `self.setNewSolution()` generate new solution and set it into self.solution
-        """
-        for self.trial_ix in range(1, int(self.n_trial) + 1):
-            # generate new solution and set it into self.solution
-            self.solution.setRandom()
+    def search(self, solution, *args):
+        for _ in range(int(self.n_trial)):
+            # generate new solution
+            solution.setRandom()
 
-            # if solution is better thatn incumbent, then update best solution
-            self.registerSolution(self.solution)
+            # update best solution if needed
+            self.registerSolution(solution)
 
-            # callbacks
-            for callback in self.callbacks:
-                callback([self.solution], self.best_solution, self.best_obj_value)
+            # execute callbacks
+            self.callback([solution])
 
             # check time limit
             self.raiseTimeoutIfNeeded()
