--- flopt/solvers/swarm_intelligence_searches/__init__.py
@@ -1,1 +1,0 @@
-from .shuffled_frog_leaping_search import ShuffledFrogLeapingSearch
